+++ packages/registry/src/__generated__/__manifest__.ts
@@ -1,7 +1,7 @@
 export default {
   "name": "@traversable/registry",
   "type": "module",
-  "version": "0.0.25",
+  "version": "0.0.26",
   "private": false,
   "description": "",
   "license": "MIT",